Avoid updating dynamic data properties
Manually updating dyndata property values permanently breaks instances of web parts as described in #200. This avoids saving changes to dynamic data properties from the editor. 

I do not believe they should be filtered out of the editor altogether as it can be quite useful to see a JSON representation of their values.